+++ adapters/mariadb/ops.go
@@ -13,7 +13,7 @@ import (
 
 const (
 	adapterName    = "mariadb"
-	adapterVersion = "0.1.0"
+	adapterVersion = "0.2.0"
 
 	defaultUser     = "root"
 	defaultDatabase = "probavi"